Update App.tsx
First you’ll need to add the `showGraph` property in the IState interface
defined in App.tsx. It should be of the type` boolean`
 Next, in the constructor of the App component, you should define that the
initial state of the App not to show the graph yet. This is because we want the
graph to show when the user clicks ‘Start Streaming Data’. That means you
should set `showGraph` property of the App’s state to `false` in the
constructor
To ensure that the graph doesn’t render until a user clicks the ‘Start
Streaming’ button, you should also edit the `renderGraph` method of the
App. In there, you must add a condition to only render the graph when the
state’s `showGraph` property of the App’s state is `true`.