Update Graph.tsx
First, you must enable the `PerspectiveViewerElement` to behave like an
HTMLElement. To do this, you can extend the `HTMLElement` class from the`PerspectiveViewerElement` interface.
Since you’ve changed the `PerspectiveViewerElement` to extend the
`HTMLElement` earlier, you can now make the definition of the `const elem`
simpler, i.e. you can just assign it straight to the result of the
`document.getElementsByTagName`. 
Finally, you need to add more attributes to the element. For this you have to
have read thru the Perspective configurations particularly on the table.view
configurations. You’ll need to add the following attributes: `view`,
`column-pivots`, `row-pivots`, `columns` and `aggregates` . If you
remember the earlier observations we did in the earlier slides, this is the
configurable part of the table/graph.